requirements.txt
sqlparse == 0.1.19
Markdown == 2.6.4
twitter == 1.17.1
selenium == 2.48.0
djangorestframework == 3.3.1
djangorestframework-filters==0.9.1
Django == 1.8.6
Scrapy == 1.0.3
Pygments == 2.0.2
django_appconf == 1.0.1
mechanize == 0.2.5
beautifulsoup4 == 4.4.1
python-creole == 1.3.1
requests == 2.8.1
hurry.filesize == 0.9
Pillow==4.2.1
